changed the url to /barcode/add and made separate components for each section
