+++ src/lib/api/product.ts
@@ -220,7 +220,8 @@ export type SelectedImage = {
 	y2: string;
 };
 
-type RawImage = {
+export type RawImage = {
+	url: '';
 	sizes: {
 		full: ImageSize;
 		100: ImageSize;
