+++ src/routes/products/search/+page.svelte
@@ -28,13 +28,13 @@
 				<SmallProductCard {product} />
 			{/each}
 		{:else}
-			<div class="mt-8 text-center opacity-70">
+			<div class="mt-8 text-center">
 				<p>No products found</p>
 				<p>We couldn't find any products matching your search</p>
-        <a
-				class="btn btn-secondary join-item mt-5 px-10"
-				href="/addProduct?barcode={page.url.searchParams.get('q')}">Add new Product</a
-			  >
+				<a
+					class="btn btn-secondary join-item mt-5 px-10"
+					href="/products/{page.url.searchParams.get('q')}/add">Add new Product</a
+				>
 			</div>
 		{/if}
 	{/await}
